+++ packages/dough_sensors/lib/dough_sensors.dart
@@ -1,7 +1,1 @@
-library dough_sensors;
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/gyro_dough.dart';