+++ packages/dough_sensors/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Extracted gyro dough from dough into its own package